refactor: merge inference speedup

+++ src/neuron_proofreader/machine_learning/image_dataloader.py
@@ -175,7 +175,7 @@ class PatchLoader(ABC):
         subgraph = self.graph.rooted_subgraph(node, depth)
 
         # Annotate mask
-        mask = np.zeros(shape)
+        mask = np.zeros(shape, dtype=np.float32)
         #self.annotate_foreground(mask, nodes, offset, fill_val=0.5)  TEMP
         self.annotate_fragment(mask, subgraph, offset, fill_val=1)
         return mask
@@ -199,7 +199,7 @@ class PatchLoader(ABC):
         patch = self.img.read(center, shape)
         patch = np.minimum(patch, self.brightness_clip)
         patch = img_util.normalize(patch, percentiles=self.percentiles)
-        return patch
+        return patch.astype(np.float32)
 
     # --- Helpers ---
     def __getattr__(self, name):
